Change the variables type to long

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	int fib1 = 0, fib2 = 1, sum = 0;
+	long int fib1 = 0, fib2 = 1, sum = 0;
 	int count;
 
 	for (count = 0; count < 50; count++)
